Improve test to cover more cases
