+++ src/test/java/net/openhft/chronicle/bytes/HeapByteStoreTest.java
@@ -21,11 +21,6 @@ import net.openhft.chronicle.bytes.internal.HeapBytesStore;
 import org.jetbrains.annotations.NotNull;
 import org.junit.Test;
 
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
-import java.nio.charset.StandardCharsets;
-import java.util.concurrent.Callable;
-
 import static org.junit.Assert.*;
 
 public class HeapByteStoreTest extends BytesTestCommon {
@@ -52,36 +47,5 @@ public class HeapByteStoreTest extends BytesTestCommon {
         // ensure this succeeds even though we are above the real capacity - this should trigger resize
         bytes.prewriteInt(1);
         assertTrue(bytes.realCapacity()> initialCapacity);
-    }
-
-    @Test
-    public void testWriteOnHeap() throws Exception {
-        doTestWrite(() -> ByteBuffer.allocate(128));
-    }
-
-    @Test
-    public void testWriteDirect() throws Exception {
-        doTestWrite(() -> ByteBuffer.allocateDirect(128));
-    }
-
-    private void doTestWrite(Callable<ByteBuffer> generator) throws Exception {
-        final Bytes<?> data = Bytes.elasticHeapByteBuffer(128);
-        ByteBuffer buffer = generator.call();
-
-        for (byte c = ' '; c < '`'; c++) {
-            data.writeChar((char)c);
-            buffer.put(c);
-        }
-
-        HeapBytesStore heapBytesStore = (HeapBytesStore) data.bytesStore();
-        heapBytesStore.write(16, buffer, 32, 8);
-        for (int i = 0; i < 16; i++)
-            assertEquals(i + ' ', heapBytesStore.readByte(i));
-
-        for (int i = 16; i < 24; i++)
-            assertEquals(i + '0', heapBytesStore.readByte(i));
-
-        for (int i = 24; i < 32; i++)
-            assertEquals(i + ' ', heapBytesStore.readByte(i));
     }
 }
